requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
Faker==38.0.0
fuzzywuzzy==0.18.0
idna==3.11
Levenshtein==0.27.3
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
scipy==1.16.3
simsimd==6.5.3
six==1.17.0
sqlparse==0.5.3
stringzilla==4.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0